backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=25.1.0",
    "bcrypt>=5.0.0",
    "fastapi>=0.135.3",
    "litellm>=1.83.2",
    "python-dotenv>=1.0.1",
    "python-jose[cryptography]>=3.3.0",
    "sqlmodel>=0.0.38",
    "uvicorn[standard]>=0.43.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
]
